hardware: Adding initial hardware
This is the hardware for the "MonkeyBadge", an as yet to be released
"BadgeLife" device.  This device features an Si4702 FM radio reciever,
an SSD1306 OLED screen, infrared TX/RX, to "standardized add-on" (SAO)
ports, and other goodies.

This spin of the hardware went to initial manufacturing with the gerbers
provided on October 22nd, 2022.