hw: MicroPython board definition files.
These files are used to compile MicroPython with the specific
configuration for the Hushcon West 2023 "MonkeyBadge".

Within MicroPython there are two primary environment variables which can
be used to configure the project:

  - `MICROPYTHON_BOARD`
  - `MICROPYTHON_BOARD_DIR`

In our case we set `MICROPYTHON_BOARD` to "monkeybadge" and
`MICROPYTHON_BOARD_DIR` to the path of this directory.

This file is used by the CMake build system for configuration.