chore: Adding a build action for the API server
This is a basic scaffolding of the use of GitHub actions to monitor the
repository for changes to the API server code.  In the event that there
is a change to the API server and the branch is merged into the main
branch, this action will build the API server and push the resulting
image to the GitHub container registry (ghcr.io).

This action is triggered by a push to the main branch.

Additionally, only have images in github.com/packetchat/mb-api-server
tagged when performing merges into the main branch.