Feature: Adding a self-hosted runner
Adding a self-hosting runner for the purpose of the local build system.

We're not worried with overloading the system, so (for now) taking off
some of the restrictions and allowing any changes to the GitHub
workflows to trigger a build.

Previously using the basic GitHub action was taking over five minutes to
complete.  I'm hoping this gets us under 30 seconds.  :fingers_crossed:

This cleans up some of the annotations in the workflow definition file
including making it clear that it's running on the local server used for
the conference.