bug: SI470X method rename (set_volume)
In a prior update to the MonkeyBadge class the method set_volume was
renamed to setVolume().  This was discovered when performing a manual
audit of the code base, as can be seen below:

```
>>> b.radio.changeVolume(12)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "library/radio.py", line 551, in changeVolume
AttributeError: 'SI470X' object has no attribute 'set_volume'
```

This commit both resolves the naming issue for the function of the class
as well as removing an extraneous carriage return at the end of the
file.  As such, we no longer need the "Tuning" property on the class nor
the missing `tune()` function as called.