chore: cleanup some minor output messages
db was outputting an ugly error on a failed get. I commented this out for now.
- TODO: add a debug warn for this

konomi - formatting

monkeybadge - added back the ip/ticks/now output but commented it out for now.
- TODO: add a debug warn for this

I also added a final peroid to format the output on the REPL once an event
kicks off