feature: SHA hash of running partition
This adds a new method "check_self_sha" to the OTA object.  From there
this method is utilized when checking verifying the desired versus
expected firmware state as expressed via JSON metadata.

If the hash of the running firmware matches the hash as provided by the
remote server there's not even a point in downloading the binary file.