feature: Battery Meter (#48)
Adding a class for measuring the battery life. The `Meter` class
contains an `info()` method that calculates the battery life based on the
measurement returned of the ADC reading. The math is a little weird
because I had to average it on various batteries and there's a voltage
divider that gives me only 2.4V to measure.

Changes:
- Constants were added to `config.py`.
- The battery class was added to `library/`.
- A helper function `battery_check` is used to call and display the
  estimated remaining battery life.