feature: OLED brightness controls
Adding a menu item with a helper function to set the OLED brightness
globally. These settings are contrast values ranging from 1-255. Three
settings exist for setting the contrast to low, medium and high which
set it to 1, 80 and 200 respectively.

I also added a default contrast setting of 1, so that when the badge
boots the brightness starts at the lowest possible value to preserve
battery.

The contrast mode is set from the menu via a lambda function so that it
doesn't set the contrast when the menu is activated. It calls a helper
function that has a parmeter called `mode` that you pass. The `mode`
parmaeter can be one of the three settings: `low`, `medium` and `high`.

Files modified:

- library/display.py
- library/monkeybadge.py