feature: Added all LED light shows
Added a dictionary `lightshow_dict` to `library/leds.py` to store a key
value pair for each light show function and a friendly name. Then I
initialized an empty list `selected_lightshows` to store 5 random
light shows to be displayed in the menu. Added a helper function
`selected_lightshows()` that returns the list. I also refactored any
light show that took parameters and internalized all the values within
each light show method. The only method that I didn't overhaul was the
normal `do_monkey_spaz()` method because it passed the name of the
monkey to determine the group and color that was flashed. Only way I
could think of how to do this is break it down into three separate
methods to handle each monkey face grouping for the show, but decided to
just comment the entire thing out and we can fix it later if we have
time.

Modified `library/monkeybadge.py` to add a for loop in the menu to
populate the 5 randomly selected light shows. Left the commented out
original menu item and the temporary light show menu for testing that
has all light shows in the menu so you can see if any of them work to
troubleshoot. Might need this before this gets pushed to main for
testing threading.

Speaking of threading. I had to modify `config.py` and
`library/monkeybadge.py` because the light shows were very jittery. In
`config.py` I changed the 6 second delay to 60 seconds, then changed the
sleep time at the bottom of `library/monkeybadge.py` to increase the
delay a bit and that smoothed out the light shows.

Files changed:
- config.py
- library/leds.py
- library/monkeybadge.py

Tested on the badge with all light shows and verified that they all
work. Still have an issue with the lightshows being slow intermitently,
but they're smoother than they were before the changes made for
adjusting the sleep time.