fix: Cleaning up superfluous comments
Removed all the testing code that was commented out and any unnecessary
comment blocks left over from development. There's still one block
commented out in flash_badge, but I'm not touching that for now in case
that was intensional.

Copied the code over to the badge and tested, works well.

Files changed:

- library/leds.py
- library/monkeybadge.py