Add hardware history (#66)
* hardware: Adding initial hardware

This is the hardware for the "MonkeyBadge", an as yet to be released
"BadgeLife" device.  This device features an Si4702 FM radio reciever,
an SSD1306 OLED screen, infrared TX/RX, to "standardized add-on" (SAO)
ports, and other goodies.

This spin of the hardware went to initial manufacturing with the gerbers
provided on October 22nd, 2022.

* hardware: Badge revision v2.2

After some initial testing we came up with a number of errata as noted
here.  These changes have been implemented:

- Need to replace the infrared LED emitter with one that is
  front-facing, and doesn’t require hand-soldering.
- Make a true reset button
- Replace the audio connector with a thru-hole for strength
- Tent the vias so white is white
- Drag a few more parts off of the silk
- Figure out a better way to do powered-off re-charge

This spin of the hardware went to initial manufacturing with the gerbers
provided on November 30th, 2022.

* hardware: Badge revision 3.4

After a holiday break work on revision 3.4 of the Monkey badge began
March 24th, 2023.

In this revision changes were made to the following components:
- power regulator
- Charge/powerpath controller
- the headphone jack
- Resistor/capacitor values on the LM4808 amplifer
- Removal of experimental touch pads

This spin of the hardware went to initial manufacturing with the gerbers
provided on July 25th, 2023.

* hardware: Badge revision 3.5

In this revision changes were made to the following components:

- Stereo amplifier
- Microphone amplifier
- Addition of a JTAG port

This spin of the hardware went to initial manufacturing with the gerbers
provided on August 30th, 2023.

* hardware: Badge revision 3.6

In this revision changes were made to the following components:

- Change C42 in stereo amplifier from 390nF to 470nF
- Changed JP4 default state (Microphone amplifier - enable amplifier)

This spin of the hardware went to initial manufacturing with the gerbers
provided on September 9th, 2023.

* hardware: Badge revision 3.7

In this revision changes were made to the following components:

- Change C42 in stereo amplifier from 470nF to 390nF
- Changed JP4 default state (Microphone amplifier - disabled amp by
  default)

This spin of the hardware went to initial manufacturing with the gerbers
provided on September 15th, 2023.

* hardware: rename "hardware" to "pcb_design"

This renames the attached hardware definitions to `pcb_design` making
it clear that these are the files related to the KiCad components and
not per-se Wokwi or anything else hardware related.

---------

Co-authored-by: Temtel <temtel@gmail.com>