bug: if a badge is trying to checkin and the server has lost the data, the badge is SOL
In the situation where a badge's DB key is lost, the badge is unusable.

to fix this, I added a check for a 404 from checkin on the apiserver,
if the badge receives a 404 - that means the badge doesnt exist.

the badge will try to re-register with the current API token. If the token exists in the api-keys
table, it's ignored, otherwise it will be added as well.

I also changed the initial last_checkin value to -61000 to ensure we checkin on the first event loop.

forgot to remove an err