feature: Battery charge warning message
Added a small power check in the main controller. I figured this is
where you should put this since it's so important.

The power level is first stored, then sliced and the whole integer is
used as the threshold, currently set to `5`. The power meter's range is
definitely safe enough to go lower, but just in case some batteries
quickly discharge when they get close to 0% I set it to 5%. This can be
changed, just thinking it could stay there instead of wait till it hits
0. I'm open to suggestions.

You can test by setting the `MIN_POWER_READING' in `config.py` to like
95 or something, ensure the badge unpluged and press the reset button;
you'll see the `show_timed_alert` message. This message is in the
controller loop so it will pop the message continuously forcing the user
to power off the badge and plug in.

There might be a problem because they'll have to charge when they turn
it back on, but this forces them to not destroy the battery, so yeah, I
think that was the right move. Open to suggestions, though. I tried to
keep the changes to the main loop to a minimum.